Separate panic handlers out by operating system. On windows also log panic to event viewer.
