+++ .gitignore
@@ -14,3 +14,6 @@ bin/
 *.exe
 *.syso
 *.log
+
+#file generated at build time by message compiler
+agentmsg.h
