+++ cmd/agent/.gitignore
@@ -1,0 +1,3 @@
+# files generated by message compiler
+agentmsg.rc
+*.bin
