+++ cmd/agent/agent.rc
@@ -46,3 +46,9 @@ BEGIN
         VALUE "Translation", 0x409, 1200
     END
 END
+
+// Country: United States
+// Language: English
+#pragma code_page(437)
+LANGUAGE 0x9, 0x1
+1 MESSAGETABLE "MSG00409.bin"
