+++ cmd/agent/agentmsg.mc
@@ -1,0 +1,16 @@
+;// Header
+MessageIdTypedef=DWORD
+
+LanguageNames=(
+    English=0x409:MSG00409
+)
+
+
+;// Messages
+MessageId=1
+SymbolicName=API_ERROR
+Severity=Error
+Language=English
+The datadog agent process unexpectedly quit.  The error was %1.
+.
+
