+++ omnibus/resources/datadog-agent6/msi/source.wxs.erb
@@ -170,8 +170,13 @@
                           Remove="uninstall"
                           Wait="yes" />
           <RemoveFolder Id="APPLICATIONROOTDIRECTORY" On="uninstall" />
-          
+          <util:EventSource Log="Application" Name="Datadog Agent"
+                EventMessageFile="[DIST]agent.exe"
+                SupportsErrors="yes"
+                SupportsInformationals="yes"
+                SupportsWarnings="yes"/>
         </Component>
+
       </Directory>
     </DirectoryRef>
